Added blog auth guards

+++ app/Http/Controllers/BlogPostsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\BlogPost;
 
 class BlogPostsController extends Controller
@@ -15,7 +16,10 @@ class BlogPostsController extends Controller
      */
     public function index()
     {
+        // Get all posts
         $posts = BlogPost::orderBy('id', 'DESC')->paginate(5);
+
+        // Return results with pagination, give the header title
         return view('blog.allPostsWithPagination')->with([
             'posts' => $posts,
             'headerType' => 'viewAll'
@@ -29,6 +33,12 @@ class BlogPostsController extends Controller
      */
     public function create()
     {
+        // Verify that the user is authenticated
+        if ( ! Auth::check() ) {
+            abort(401);
+        }
+
+        // Show the view
         return view('blog.createPost');
     }
 
@@ -40,18 +50,25 @@ class BlogPostsController extends Controller
      */
     public function store(Request $request)
     {
+        // Verify that the user is authenticated
+        if ( ! Auth::check() ) {
+            abort(401);
+        }
+
+        // Form validation
         $request->validate([
             'title' => 'required',
             'body' => 'required'
         ]);
 
-        $input = $request->all();
+        // Create the blog post
         $id = BlogPost::create([
-            'author_id' => 1,
-            'title'     => $input['title'],
-            'body'      => $input['body']
+            'author_id' => Auth::id(),
+            'title'     => $request->input('title'),
+            'body'      => $request->input('body')
         ])->id;
 
+        // Show the created blog post view
         return redirect('/blog/' . $id);
     }
 
@@ -63,7 +80,10 @@ class BlogPostsController extends Controller
      */
     public function show($id)
     {
+        // Retrieve the requested blog post
         $post = BlogPost::findOrFail($id);
+
+        // Show the view with the blog post
         return view('blog.viewPost')->with('post', $post);
     }
 
@@ -75,7 +95,15 @@ class BlogPostsController extends Controller
      */
     public function edit($id)
     {
+        // Verify that the user is authenticated
+        if ( ! Auth::check() ) {
+            abort(401);
+        }
+
+        // Retrieve the requested blog post
         $post = BlogPost::findOrFail($id);
+
+        // Show the view with the blog post
         return view('blog.editPost')->with('post', $post);
     }
 
@@ -88,18 +116,28 @@ class BlogPostsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // Verify that the user is authenticated
+        if ( ! Auth::check() ) {
+            abort(401);
+        }
+
+        // Validate the form
         $request->validate([
             'title' => 'required',
             'body' => 'required'
         ]);
 
+        // Find the requested blog post
         $post = BlogPost::findOrFail($id);
+
+        // Update the blog post
         $post->update([
             'author_id' => 1,
             'title'     => $request->input('title'),
             'body'      => $request->input('body')
         ]);
 
+        // Show the view with the created blog post
         return view('blog.viewPost')->with('post', $post);
     }
 
@@ -111,8 +149,15 @@ class BlogPostsController extends Controller
      */
     public function destroy($id)
     {
+        // Verify that the user is authenticated
+        if ( ! Auth::check() ) {
+            abort(401);
+        }
+
+        // Retrieve the requested blog post and delete it
         BlogPost::findOrFail($id)->delete();
-        $posts = BlogPost::orderBy('id', 'DESC')->paginate(5);
+
+        // Redirect to the main blog page
         return redirect('/blog');
     }
 
@@ -124,9 +169,13 @@ class BlogPostsController extends Controller
      */
     public function search(Request $request)
     {
+        // Store the query
         $query = $request->input('query');
+
+        // Search the database for posts matching the query
         $posts = BlogPost::where('body', 'LIKE', "%{$query}%")->paginate(5);
         
+        // Return the view with the matching blog posts
         return view('blog.allPostsWithPagination')->with([
             'posts' => $posts, 
             'headerType' => 'search',
